Seperate main function and Inner classes from StandaloneGameOfLife to each classes
