+++ src/org/bitstorm/gameoflife/ManualMenuItemListener.java
@@ -1,0 +1,19 @@
+package org.bitstorm.gameoflife;
+
+import org.bitstorm.util.TextFileDialog;
+
+import java.awt.*;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+
+public class ManualMenuItemListener implements ActionListener {
+	private Frame frame;
+	public ManualMenuItemListener(Frame frame){
+		this.frame = frame;
+	}
+	@Override
+	public synchronized void actionPerformed(ActionEvent e) {
+		Point p = frame.getLocation();
+		new TextFileDialog(frame, "Game of Life License", "license.txt", p.x + 60, p.y + 60);
+	}
+}
