+++ src/org/bitstorm/gameoflife/eventhandler/DragDrawHandler.java
@@ -1,14 +1,15 @@
 package org.bitstorm.gameoflife.eventhandler;
 
+import org.bitstorm.gameoflife.ui.CellGridDrawer;
 import org.bitstorm.gameoflife.ui.GameOfLifeAWTCellGrid;
 
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseMotionListener;
 
-public class CellGridCanvasMouseMotionHandler implements MouseMotionListener {
-	private GameOfLifeAWTCellGrid canvas;
+public class DragDrawHandler implements MouseMotionListener {
+	private CellGridDrawer canvas;
 	
-	public CellGridCanvasMouseMotionHandler(GameOfLifeAWTCellGrid canvas){
+	public DragDrawHandler(CellGridDrawer canvas){
 		this.canvas = canvas;
 	}
